renamed plugins to connection
