+++ kedro_mlflow/config/__init__.py
@@ -1,1 +1,1 @@
-from .plugin import KedroMlflowConnection
+from .connection import KedroMlflowConnection
