+++ tests/config/test_kedro_mlflow_connection.py
@@ -1,6 +1,6 @@
 import pytest
 
-from kedro_mlflow.config.plugin import KedroMlflowConnection
+from kedro_mlflow.config.connection import KedroMlflowConnection
 
 
 class FakeConnection(KedroMlflowConnection):
